pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seeka"
version = "0.1.0"
description = "Embedded Agent memory component"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "pyseekdb",
    "sentence-transformers",
    "chakpy",
    "aiosqlite",
    "openai>=1.0",
    "cohere>=5.0",
    "python-dotenv",
    "nanoid",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["seeka*"]
